change return type of ContestManagerBean.getContestPostCount from BigDecimal to int.
git-svn-id: https://coder.topcoder.com/tcs/clients/cronos/components@132303 fb370eea-3af6-4597-97f7-f7400a59c12a

+++ studio_contest_manager/trunk/src/java/main/com/topcoder/service/studio/contest/ContestManager.java
@@ -936,7 +936,7 @@ public interface ContestManager {
      * @throws ContestManagementException
      *             if any error occurs when getting contest post count.
      */
-    public BigDecimal getContestPostCount(long forumId) throws ContestManagementException;
+    public int getContestPostCount(long forumId) throws ContestManagementException;
 
     /**
      * <p>
